requirements.txt
annotated-types==0.7.0
assertpy==1.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
curlify==2.2.1
idna==3.10
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pydentic==0.0.1.dev3
PyHamcrest==2.1.0
pytest==8.3.5
python-stdnum==1.20
requests==2.32.3
retrying==1.3.4
six==1.17.0
structlog==25.1.0
typing_extensions==4.12.2
urllib3==2.3.0
